+++ R/proj.R
@@ -36,15 +36,8 @@ to_proj <- function(x) {
 
   ## integer of 4 or 5 digits,
   ## or is a character string
-  if (is.numeric(x) || (nchar(x) %in% c(4, 5, 6) && grepl("^[0-9]{1,5}$", x))) {
-    ## here we need PROJ::ok_proj6() pivot
-    if (!.ok_PROJ()) {
-      ## we are PROJ library version < 6
-      x <- sprintf("+init=epsg:%i", as.integer(x))
-    } else {
-      ## we are PROJ library version >= 6
-      x <- sprintf("EPSG:%i", as.integer(x))
-    }
+  if (is.numeric(x) || (nchar(x) %in% c(4, 5, 6) && grepl("^[0-9]{1,6}$", x))) {
+    x <- sprintf("EPSG:%i", as.integer(x))
   }
   x <- trimws(x, which = "left")
   ## fix the weird sf/raster thing where raster::projection() is just the datum string