+++ R/reproj-package.R
@@ -7,8 +7,8 @@ NULL
 
 #' Reproject data from source to target coordinate system.
 #'
-#' reproj provides helpers for easily reprojecting generic data, by depending
-#' on a reprojection engine (proj4 for now).
+#' reproj provides helpers for easily reprojecting generic data, using the
+#' PROJ package as the reprojection engine.
 #'
 #' The function `reproj` is designed to take an input data set `x` and then a
 #' `target` coordinate system specification. The `source` argument is not positional
@@ -22,10 +22,5 @@ NULL
 #'
 #' See [reproj] for global options to control assumptions about data that
 #' is input in longitude latitude form.
-#'
-#' There is an option set at start up `reproj.mock.noproj6` which is designed
-#' for testing the support in the PROJ package. Even if this package is
-#' functional this option can be set to true so that reproj falls-back to
-#' use the proj4 package instead.
 #' @name reproj-package
 NULL